spike: moving twowaybinding file; pr feedback
