+++ packages/plugin-core/src/components/lookup/NoteLookupProvider.ts
@@ -73,7 +73,9 @@ export class NoteLookupProvider implements ILookupProviderV3 {
       },
       100,
       {
-        leading: true,
+        // Use trailing to make sure we get the latest letters typed by the user
+        // before accepting.
+        leading: false,
       }
     );
     quickpick.onDidChangeValue(onUpdateDebounced);
@@ -83,7 +85,7 @@ export class NoteLookupProvider implements ILookupProviderV3 {
         ctx: "NoteLookupProvider:onDidAccept",
         quickpick: quickpick.value,
       });
-      onUpdateDebounced.cancel();
+      await onUpdateDebounced.flush();
       if (_.isEmpty(quickpick.selectedItems)) {
         await onUpdatePickerItems({
           picker: quickpick,
