+++ packages/plugin-core/src/test/suite-integ/components/lookup/LookupControllerV3.test.ts
@@ -21,7 +21,7 @@ import {
 import { LookupControllerV3 } from "../../../../../src/components/lookup/LookupControllerV3";
 import { NoteLookupProviderFactory } from "../../../../../src/components/lookup/LookupProviderV3Factory";
 import { ExtensionProvider } from "../../../../../src/ExtensionProvider";
-import { TwoWayBinding } from "../../../../../src/types/TwoWayBinding";
+import { TwoWayBinding } from "../../../../utils/TwoWayBinding";
 import { WSUtilsV2 } from "../../../../../src/WSUtilsV2";
 import { VaultSelectionMode } from "../../../../components/lookup/types";
 import { expect } from "../../../testUtilsv2";
